Build and deploy app as a .deb package
Also configure it so it doesn't write to files in the deployed /usr/
directory:

  - log to stdout instead of log/<environment>.log
  - cache files in a dedicated data dir instead of tmp/cache
  - read secret_key_base from an ENV Var instead of auto-generating

These weren't strictly necessary, but I wanted the code deployed to
/usr/local/share/<application> to be read-only. Some of these changes
would also be good to remember when deploying to Docker (eg. logging).

Also deploy Nginx Unit to run the deployed application. The missing bit
of this is managing the environment for running `rake`/`rails` commands
such as `db:migrate`: we'll need a way to populate the same variables
into a shell to run those commands. Is there a way to avoid duplication?